ncrypt: refactor sort_*.c
Factor out the GPGME and PGP Key sort functions:

- ncrypt/sort_gpgme.c
- ncrypt/sort_pgp.c
