+++ ncrypt/dlg_gpgme.c
@@ -104,135 +104,6 @@ static const struct Mapping GpgmeHelp[] = {
 };
 
 /**
- * crypt_sort_address - Compare two keys by their addresses - Implements ::sort_t - @ingroup sort_api
- */
-static int crypt_sort_address(const void *a, const void *b, void *sdata)
-{
-  struct CryptKeyInfo *s = *(struct CryptKeyInfo **) a;
-  struct CryptKeyInfo *t = *(struct CryptKeyInfo **) b;
-  const bool sort_reverse = *(bool *) sdata;
-
-  int rc = mutt_istr_cmp(s->uid, t->uid);
-  if (rc != 0)
-    goto done;
-
-  rc = mutt_istr_cmp(crypt_fpr_or_lkeyid(s), crypt_fpr_or_lkeyid(t));
-
-done:
-  return sort_reverse ? -rc : rc;
-}
-
-/**
- * crypt_sort_keyid - Compare two keys by their IDs - Implements ::sort_t - @ingroup sort_api
- */
-static int crypt_sort_keyid(const void *a, const void *b, void *sdata)
-{
-  struct CryptKeyInfo *s = *(struct CryptKeyInfo **) a;
-  struct CryptKeyInfo *t = *(struct CryptKeyInfo **) b;
-  const bool sort_reverse = *(bool *) sdata;
-
-  int rc = mutt_istr_cmp(crypt_fpr_or_lkeyid(s), crypt_fpr_or_lkeyid(t));
-  if (rc != 0)
-    goto done;
-
-  rc = mutt_istr_cmp(s->uid, t->uid);
-
-done:
-  return sort_reverse ? -rc : rc;
-}
-
-/**
- * crypt_sort_date - Compare two keys by their dates - Implements ::sort_t - @ingroup sort_api
- */
-static int crypt_sort_date(const void *a, const void *b, void *sdata)
-{
-  struct CryptKeyInfo *s = *(struct CryptKeyInfo **) a;
-  struct CryptKeyInfo *t = *(struct CryptKeyInfo **) b;
-  const bool sort_reverse = *(bool *) sdata;
-
-  unsigned long ts = 0;
-  unsigned long tt = 0;
-  int rc = 0;
-
-  if (s->kobj->subkeys && (s->kobj->subkeys->timestamp > 0))
-    ts = s->kobj->subkeys->timestamp;
-  if (t->kobj->subkeys && (t->kobj->subkeys->timestamp > 0))
-    tt = t->kobj->subkeys->timestamp;
-
-  if (ts > tt)
-  {
-    rc = 1;
-    goto done;
-  }
-
-  if (ts < tt)
-  {
-    rc = -1;
-    goto done;
-  }
-
-  rc = mutt_istr_cmp(s->uid, t->uid);
-
-done:
-  return sort_reverse ? -rc : rc;
-}
-
-/**
- * crypt_sort_trust - Compare two keys by their trust levels - Implements ::sort_t - @ingroup sort_api
- */
-static int crypt_sort_trust(const void *a, const void *b, void *sdata)
-{
-  struct CryptKeyInfo *s = *(struct CryptKeyInfo **) a;
-  struct CryptKeyInfo *t = *(struct CryptKeyInfo **) b;
-  const bool sort_reverse = *(bool *) sdata;
-
-  unsigned long ts = 0;
-  unsigned long tt = 0;
-
-  int rc = mutt_numeric_cmp(s->flags & KEYFLAG_RESTRICTIONS, t->flags & KEYFLAG_RESTRICTIONS);
-  if (rc != 0)
-    goto done;
-
-  // Note: reversed
-  rc = mutt_numeric_cmp(t->validity, s->validity);
-  if (rc != 0)
-    return rc;
-
-  ts = 0;
-  tt = 0;
-  if (s->kobj->subkeys)
-    ts = s->kobj->subkeys->length;
-  if (t->kobj->subkeys)
-    tt = t->kobj->subkeys->length;
-
-  // Note: reversed
-  rc = mutt_numeric_cmp(tt, ts);
-  if (rc != 0)
-    goto done;
-
-  ts = 0;
-  tt = 0;
-  if (s->kobj->subkeys && (s->kobj->subkeys->timestamp > 0))
-    ts = s->kobj->subkeys->timestamp;
-  if (t->kobj->subkeys && (t->kobj->subkeys->timestamp > 0))
-    tt = t->kobj->subkeys->timestamp;
-
-  // Note: reversed
-  rc = mutt_numeric_cmp(tt, ts);
-  if (rc != 0)
-    goto done;
-
-  rc = mutt_istr_cmp(s->uid, t->uid);
-  if (rc != 0)
-    goto done;
-
-  rc = mutt_istr_cmp(crypt_fpr_or_lkeyid(s), crypt_fpr_or_lkeyid(t));
-
-done:
-  return sort_reverse ? -rc : rc;
-}
-
-/**
  * crypt_key_abilities - Parse key flags into a string
  * @param flags Flags, see #KeyFlags
  * @retval ptr Flag string
@@ -642,30 +513,7 @@ struct CryptKeyInfo *dlg_gpgme(struct CryptKeyInfo *keys, struct Address *p,
     return NULL;
   }
 
-  const short c_pgp_sort_keys = cs_subset_sort(NeoMutt->sub, "pgp_sort_keys");
-  sort_t fn = NULL;
-  switch (c_pgp_sort_keys & SORT_MASK)
-  {
-    case KEY_SORT_ADDRESS:
-      fn = crypt_sort_address;
-      break;
-    case KEY_SORT_DATE:
-      fn = crypt_sort_date;
-      break;
-    case KEY_SORT_KEYID:
-      fn = crypt_sort_keyid;
-      break;
-    case KEY_SORT_TRUST:
-    default:
-      fn = crypt_sort_trust;
-      break;
-  }
-
-  if (ARRAY_SIZE(&ckia) > 1)
-  {
-    bool sort_reverse = c_pgp_sort_keys & SORT_REVERSE;
-    ARRAY_SORT(&ckia, fn, &sort_reverse);
-  }
+  gpgme_sort_keys(&ckia);
 
   enum MenuType menu_to_use = MENU_GENERIC;
   if (app & APPLICATION_PGP)
